Show date on PDF
