+++ packages/web-frontend/src/utils/periodGranularities.js
@@ -338,3 +338,34 @@ export const constrainDate = (date, min, max) => {
   if (d > max) d = max;
   return d;
 };
+
+export const getDatesAsString = (isSingleDate, granularity, startDate, endDate) => {
+  const DEFAULT_GRANULARITY = GRANULARITY_CONFIG[GRANULARITIES.DAY];
+  const { rangeFormat } = GRANULARITY_CONFIG[granularity] || DEFAULT_GRANULARITY;
+
+  const formattedStartDate = momentToDateString(startDate, granularity, rangeFormat);
+  const formattedEndDate = momentToDateString(endDate, granularity, rangeFormat);
+
+  return isSingleDate ? formattedEndDate : `${formattedStartDate} - ${formattedEndDate}`;
+};
+
+export const getDefaultStartDateAndEndDate = (
+  isSingleDate,
+  granularity,
+  minMomentDate = moment(DEFAULT_MIN_DATE),
+  maxMomentDate = moment(),
+) => {
+  let defaultStartDate;
+  let defaultEndDate;
+  if (isSingleDate) {
+    defaultStartDate = constrainDate(moment(), minMomentDate, maxMomentDate);
+    defaultEndDate = defaultStartDate; // end date is the same, but gets rounded to the period below
+  } else {
+    defaultStartDate = minMomentDate;
+    defaultEndDate = maxMomentDate;
+  }
+  const roundedDefaults = roundStartEndDates(granularity, defaultStartDate, defaultEndDate);
+  defaultStartDate = roundedDefaults.startDate;
+  defaultEndDate = roundedDefaults.endDate;
+  return { defaultStartDate, defaultEndDate };
+};
